requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
Flask-Migrate==4.0.5
jsonschema==4.22.0
lxml==5.2.1 